backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.14
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
email-validator==2.1.0
# Optional database drivers (install separately if needed):
# psycopg2-binary==2.9.9  # For PostgreSQL
# pymysql==1.1.0         # For MySQL 
bcrypt 